Very bad commit for the presentation.
This is a very bad commit that I would not do in real life. I made the code not
compile to make a point. I am sorry.

Change the properties of the Profile model to all be constants. Which breaks
everything. Trust me, I was just showing them it would break. Please help.